requirements.txt
google-generativeai==0.7.1
python-dotenv==1.0.1
streamlit==1.36.0